Delete new validator base class entirely, move functionality into static helper class (no remaining instanced data)
